Move DeployDiff out of fab package
