+++ src/commcare_cloud/commands/deploy/diff_templates/console.txt.j2
