+++ src/commcare_cloud/commands/deploy/formplayer.py
@@ -11,13 +11,13 @@ from commcare_cloud.cli_utils import ask
 from commcare_cloud.colors import color_warning, color_notice, color_summary
 from commcare_cloud.commands.ansible import ansible_playbook
 from commcare_cloud.commands.ansible.helpers import AnsibleContext
+from commcare_cloud.commands.deploy.deploy_diff import DeployDiff
 from commcare_cloud.commands.deploy.sentry import update_sentry_post_deploy
 from commcare_cloud.commands.deploy.utils import record_deploy_start, record_deploy_failed, \
     announce_deploy_success, create_release_tag, DeployContext
 from commcare_cloud.user_utils import get_default_username
 from commcare_cloud.commands.utils import timeago
 from commcare_cloud.events import publish_deploy_event
-from commcare_cloud.fab.deploy_diff import DeployDiff
 from commcare_cloud.github import github_repo
 
 AWS_BASE_URL_ENV = {
